+++ src/decode/image_decoder.hpp
@@ -22,7 +22,7 @@
 #include "./bmp_decoder.hpp"
 #include "./tiff_decoder.hpp"
 #include "../core/sample.hpp"
-#include "../core/object_pool.hpp"
+#include "../core/buffer_pool.hpp"
 #include "../core/compat.hpp"  // span polyfill for C++17
 #include <cstddef>
 #include <cstdint>