+++ src/decode/jpeg_decoder.hpp
@@ -13,7 +13,7 @@
 #pragma once
 
 #include "../core/sample.hpp"
-#include "../core/object_pool.hpp"
+#include "../core/buffer_pool.hpp"
 #include "../core/compat.hpp"  // span polyfill for C++17
 #include <jpeglib.h>
 #include <csetjmp>