+++ src/decode/png_decoder.hpp
@@ -15,7 +15,7 @@
 #pragma once
 
 #include "../core/sample.hpp"
-#include "../core/object_pool.hpp"
+#include "../core/buffer_pool.hpp"
 #include "../core/compat.hpp"  // span polyfill for C++17
 #include <png.h>
 #include <csetjmp>