+++ src/decode/tiff_decoder.hpp
@@ -19,7 +19,7 @@
 #ifdef HAVE_TIFF
 
 #include "../core/sample.hpp"
-#include "../core/object_pool.hpp"
+#include "../core/buffer_pool.hpp"
 #include "../core/compat.hpp"  // span polyfill for C++17
 #include <tiffio.h>
 #include <cstddef>